Focus new space after switching spaces
After switching to a new space, focus on a window in the new space.
Either focus on a provided window, or find the first visible window on
the new space. Sometimes MacOS will focus on a different window for the
same application so try three times to focus the window by first
activating the windows application.

Move the mouse to the center of the new space's screen. If there are no
windows on the new space, it is safe to click the mouse to focus the
space.