WIP ignore tabs in windows
A window with no tabs will have a tabCount of 0. A new tab for an
existing window will have a tabCount equal to the total number
of tabs (eg. 2 for the first new tab). The tabCount for existing
tabs will be set back to 0. This means we need to capture and
track each new tab when it is created. We cannot identify a tab
later.

Create a getFocusedWindow() helper function.