Display tab detected notification and misc cleanup
Show a popup notification when a window with a tab is detected.
PaperWM is currently unable to handle windows with tabs. These tabbed
windows are usually Finder and Terminal, where each tab is detected as a
new hs.window. Browsers like Chrome and Safari are fine.

Include some miscellaneous code cleanup and refactoring.