More updates to space switching
Use a coroutine to handle focusing a space after switching. MacOS will
sometimes switch the focus of a window to another application when
moving a window to a new space. Try to focus a window 3 times when it is
added to a new space.

Revert to dragging a window to a new space with a mouse. Using the
hs.window:move() method fails when the old and new space are on the same
screen. We need the mouse to "pick up" a window so that it moves with
the space change.