Try to prevent MacOS from stealing focus away to another window
Often times, MacOS will steal focus away from one window to another
window within the same application. For example, focus a Chrome or
Safari window on one monitor, and the focus will quickly switch to
another visible Chrome or Safari window on a different monitor.

Start a timer to check the current window focus one animation duration
after focusing a window and try to focus the window again if the desired
window is not focused.