Add focused windows not detected by the window filter
There are some windows (like a Firefox browser window with a new
profile) that are not detected by the Hammerspoon window filter. These
windows emit no events for being opened, closed, or focused.

These windows are returned by the hs.window.focusedWindow() method if
they are the focused window.

If we attempt to tile a space and the focused window is not part of the
managed window list. Add it to the window list then. This will allow the
new window to be tiled and moved / navigated with keyboard shortcuts. It
will not fix the lack of window filter events. That needs more
investigation and maybe a Hammerspoon fix.