Fix focusing nearby window when a window is removed
When a window is closed on MacOS, another window within the same app will be focused. If there are no other windows in the app, then no window will have focus.

In PaperWM.spoon, we want to focus a nearby window (looking to the left, right, up, and down) when a window is removed.

Logs show that MacOS will sometimes send a windowFocused event (for the new window in same app) BEFORE sending the windowNotVisible and windowDestroyed events. This makes it impossible to 1) check if the removed window had focus and 2) search for nearby windows.

I think we can assume that the only way to close a window on MacOS is from user interaction when the window is focused. We'll default to always searching for a nearby window to focus. There is the skip_new_window_focus parameter that is used to avoid this behavior when we are automatically moving windows around (during refresh or moving a window to a new space).