Improve refresh windows on screen change and tiling error
Debounce screen change events by scheduling a timer to refresh windows.
Screen change events can occur multiple times in a short duration when a
monitor is plugged / unplugged, or a laptop lid is opened.

If we try to tile a window that does not have a tiling index (usually
because the window is not on a space), refresh the windows to find the
space for that window. This usually happens after a screen change or a
display size / resolution change.

These changes reduce the number of times I've had to restart the spoon.