Pick up window with lift_window modifier + left mouse drag
Use a keyboard modifier (set with PaperWM.lift_window) and a left mouse
click and drag to pick up a window (set floating) and drag it to a new
location. The window will be tiled in its new place.

This is a convenient way to reorder windows or move a window to new
space, using the mouse.