Fix and optimize slide windows
We were previously using hs.window userdata as keys for the
state.x_positions table. When stacking windows in a column, Lua decides
that the new userdata is different and creates a new entry in the
x_positions table instead of updating the window virtual x position.
Instead, use integer window IDs for x_position table keys.

Looking up a hs.window from an ID is costly. When a window slide
operation starts, cache a list of hs.windows, their virtual x positions,
and their frames. Use this cached info to set each window position from
a delta x slide update. This way we don't need to call back into MacOS
to fetch a window or any window info, and avoid a Lua -> Objective-C
bridge. There is a slight initial delay because these window lookup have
to be performed sometime, but anecdotally the window sliding behavior is
smoother overall.

Finally, if a mouse click or drag event is for a PaperWM.spoon window
movement, delete that event so it does not propagate. This prevents
MacOS from moving the window or the content of the window when the
cursor location and calculated delta x are slightly off.