Use float toggling to move window to a new space
The change to refresh windows when encountering an anchor window without
an index in tileSpace broke the moveWindowToSpace() functionality.

This is because when moving a window to a new space we first remove the
window, then tile the old space, then tile the new space. But since the
moving window is focused, it is used as the anchor window to tile the
old space. When the window index is not found the window is re-added to
the old space and moving the window to a new space has no effect.

We already have a way for a window to be tracked but not tiled:
floating. To move a window to a new space, first set it to floating,
then tile the old space (the floating window will be ignored), then move
the window to a new space, then set the window to non-floating.

Also make a few misc improvements. Have the getFirstVisibleWindow()
method return the closes off screen window if none are fully on screen.
Add more floating window helper methods to encapsulate the core floating
window state tracking.