Add helper methods for x_positions state
Make x_positions a private state variable. Create a getter method
that returns a table with accessor methods for window x positions for
a space.

Create an empty table for a space when adding a new window x position
and remove an empty table for a space when removing a window x position.

Also move the isFloatingKey into floating.lua.