Move windows to their final locations after swap / slurp / barf
I want to keep the behavior where the user sees the focused window move left or
right (and up and down) when slurping, barfing, or swaping the window. We need
to manually move some windows around to do so. The automatic tiling behavior
will keep the focused window in the same location and shift all the other
windows left or right.