Ensure new windows are placed on screen
We try our best to insert new windows to the right of the currently focused
window and make sure they are on screen. But sometimes we don't know the
previously focused window, so we insert the window into an array of tiled
windows by position on screen, from left to right.

This is better than appending a new window to the end of the tiled array, or
putting it at the beginning or in the middle, because the existing tiled windows
will not shift left or right and the new window should be on screen.

But, sometimes the center of the new window is off screen so it it's inserted
off screen. Sometimes it's not focused so it's tiled off screen.

Use setFrameInScreenBounds() to make sure the new window is on screen before
inserting it into the tiled array. Then, focus the new window after inserting so
that the space is tiled and the new window is moved on screen.

Also fix up busted window mock setFrame() and setFrameInScreenBounds().